:sparkles: Add ImagesService for avatar transformation and update user controller
